requirements.txt
streamlit
snowflake-snowpark-python
snowflake-connector-python
pandas
python-dotenv
